Move all files related to the `ViewHeader` component to a separate folder
