+++ frontend/src/metabase/query_builder/components/view/ViewHeader/View.styled.tsx
