+++ frontend/src/metabase/query_builder/components/view/ViewHeader/ViewHeaderUtils.js
